[241223_이하성] 수정: 밍글페이지-폴더명 Main => Home으로 변경, 불필요한 주석 삭제
